bot tg/requirements.txt
python-telegram-bot==22.7
Flask==2.3.3
python-dotenv==1.0.0
httpx==0.25.2
